Add basic LED and button
The code will print "blinking..." to the screen, toggle the LED on and
off a few times, then print "press me..." to the LCD waiting for someone
to press it before booting up.

Also fix precision of temp and humidity to match what the sensor
actually states it can manage. This also makes it look prettier on the
LCD.